Resolved ClangTidy issues after project inspection

+++ packages/compiler/src/ast/nodes/functions/internal_names.cpp
@@ -18,7 +18,7 @@ std::string stride::ast::resolve_internal_function_name(
     // Not perfect, but semi unique
     for (const auto& type : parameter_types)
     {
-        type_hash |= ast_type_to_internal_id(type);
+        type_hash |= static_cast<int>(ast_type_to_internal_id(type));
         type_hash <<= parameter_offset;
         parameter_offset += 2;
     }
@@ -39,7 +39,7 @@ std::string stride::ast::resolve_internal_struct_name(
     // Not perfect, but semi unique
     for (const auto& type : struct_members | std::views::values)
     {
-        type_hash |= ast_type_to_internal_id(type.get());
+        type_hash |= static_cast<int>(ast_type_to_internal_id(type.get()));
         type_hash <<= parameter_offset;
         parameter_offset += 2;
     }
